pyproject.toml
[tool.poetry]
name = "daily-coding-answers"
version = "0.1.0"
description = ""
authors = ["Darrell Calderon <dagreatssj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-api-python-client = "^2.108.0"
google-auth-httplib2 = "^0.1.1"
google-auth-oauthlib = "^1.1.0"
icecream = "^2.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
